+++ app/Exceptions/Handler.php
@@ -7,6 +7,17 @@ use Throwable;
 
 class Handler extends ExceptionHandler
 {
+
+  public function render($request, Throwable $exception)
+  {
+    // Check if the exception is an authorization exception
+    if ($exception instanceof \Illuminate\Auth\Access\AuthorizationException) {
+      return response()->view('errors.403', ['exception' => $exception], 403);
+    }
+
+    return parent::render($request, $exception);
+  }
+
     /**
      * A list of exception types with their corresponding custom log levels.
      *
@@ -47,4 +58,5 @@ class Handler extends ExceptionHandler
             //
         });
     }
+
 }