+++ database/migrations/2024_03_07_155922_add_end_time_index_to_show_schedule_table.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+      Schema::table('show_schedule', function (Blueprint $table) {
+        // Add an index to the 'end_time' column
+        $table->index('end_time', 'show_schedule_end_time_index');
+      });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+      Schema::table('show_schedule', function (Blueprint $table) {
+        // Remove the index from the 'end_time' column
+        $table->dropIndex('show_schedule_end_time_index');
+      });
+    }
+};